move Hook to registry package
