+++ server/server/session/src/main/java/com/alipay/sofa/registry/server/session/registry/ClientRegistrationHook.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.registry.server.session.strategy;
+package com.alipay.sofa.registry.server.session.registry;
 
 import com.alipay.sofa.registry.common.model.store.StoreData;
