+++ server/server/session/src/main/java/com/alipay/sofa/registry/server/session/registry/SessionRegistry.java
@@ -43,8 +43,6 @@ import com.alipay.sofa.registry.server.session.slot.SlotTableCache;
 import com.alipay.sofa.registry.server.session.store.DataStore;
 import com.alipay.sofa.registry.server.session.store.Interests;
 import com.alipay.sofa.registry.server.session.store.Watchers;
-import com.alipay.sofa.registry.server.session.strategy.ClientRegistrationHook;
-import com.alipay.sofa.registry.server.session.strategy.impl.DefaultClientRegistrationHook;
 import com.alipay.sofa.registry.server.shared.env.ServerEnv;
 import com.alipay.sofa.registry.util.ConcurrentUtils;
 import com.alipay.sofa.registry.util.LoopRunnable;
